+++ docs/changelog.md
@@ -5,6 +5,7 @@
 ### 0.0.5
 
 - Various bug fixes.
+- Added the `layer.adjusts` property, a library of tweaks and dials that allows, say, `layer.adjusts.white_balance.tint = -0.4`.
 
 ### 0.0.4 (2020-01-15)